FE Header Admim

+++ web-admin/src/pages/login_page.js
@@ -3,6 +3,7 @@ import React, { useRef } from "react";
 import { useNavigate } from "react-router-dom";
 import url_config from "../config.json";
 import "../login.css"; // Import the CSS file
+import Swal from "sweetalert2";  // Import SweetAlert2
 
 export default function LoginPage() {
   const navigation = useNavigate();
@@ -21,14 +22,41 @@ export default function LoginPage() {
       });
 
       if (status === 200) {
-        window.alert(response);
-        if (type) {
-          window.localStorage.setItem("@isAdmin", isAdmin);
-          navigation("/user", { replace: true });
+        if(response === "Đăng nhập thất bại!"){
+          Swal.fire({       
+            title: response,  // Hiển thị thông báo trả về từ server      
+            icon: 'error',  // Biểu tượng lỗi
+            confirmButtonText: 'OK',  // Nút xác nhận
+            position: 'top',  // Chỉnh vị trí xuất hiện của hộp thoại
+          }).then(() => {
+            if (type) {
+              window.localStorage.setItem("@isAdmin", isAdmin);
+              navigation("/user", { replace: true }); // Chuyển hướng sau khi đăng nhập
+            }
+          });
+        }else{
+          Swal.fire({       
+            title: response,  // Hiển thị thông báo trả về từ server      
+            icon: 'success', // Biểu tượng thành công
+            confirmButtonText: 'OK',  // Nút xác nhận
+            position: 'top',  // Chỉnh vị trí xuất hiện của hộp thoại
+          }).then(() => {
+            if (type) {
+              window.localStorage.setItem("@isAdmin", isAdmin);
+              navigation("/user", { replace: true }); // Chuyển hướng sau khi đăng nhập
+            }
+          });
         }
-      }
+        }
     } catch (error) {
-      console.log(error);
+      // Thông báo lỗi
+      Swal.fire({
+        title: 'Đăng nhập thất bại!',
+        text: 'Vui lòng kiểm tra lại thông tin đăng nhập.',
+        icon: 'error',  // Biểu tượng lỗi
+        confirmButtonText: 'Thử lại',
+        position: 'top',  // Chỉnh vị trí xuất hiện của hộp thoại
+      });
     }
   }
